requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
requests
